apps/api/pyproject.toml
[tool.poetry]
name = "shelterlink-api"
version = "1.0.0"
description = "FastAPI backend for ShelterLink - LA County shelter finder"
authors = ["ShelterLink Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-limiter = "^0.1.5"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
geoalchemy2 = "^0.14.2"
psycopg = "^3.1.13"
python-multipart = "^0.0.6"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
dev = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
test = "pytest"
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate -m"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
